modified:   lol.py  	new file:   pyingest/parsers/aps.py  	deleted:    pyingest/parsers/apsjats.py  	modified:   pyingest/parsers/jats.py
