+++ lol.py
@@ -1,13 +1,13 @@
 import glob
 import cStringIO
-import pyingest.parsers.jats as jats
+import pyingest.parsers.aps as aps
 import pyingest.serializers.classic
 
 #testfile = 'test_data/stubdata/input/apsjats_10.1103.PhysRevA.97.043801.fulltext.xml'
 testfile = glob.glob('test_data/stubdata/input/apsjats*')
 for f in testfile:
     with open(f,'rU') as fp:
-        parser = jats.JATSParser()
+        parser = aps.APSJATSParser()
         document = parser.parse(fp)
 
 #   print(document.keys())
